// repos/epaper-stock-price/src/models/interfaces.ts
export interface IUserReq {
  userId: string;
  isAdmin: boolean;
  timeZone: string;
}

export interface IUser {
  _id?: string;
  name: string;
  email: string;
  password: string;
  isActive: boolean;
  userRole: number;
  enableFastRefresh: boolean;
  timeZone: string;
  timeFormat: string;
  isAdmin?: boolean;
  token: string;
  confirmToken?: {
    token: string;
    created: Date;
  };
  resetToken?: {
    token: string;
    created: Date;
  };
  createdAt?: DateTime | string;
  updatedAt?: DateTime | string;
  noOfDevices?: number;
  meta: any;
  __v?: number;
}